fix: fix currentModel when user selected model was removed
